Rename commands_v2 to commands
