+++ src/commands/bot/admin/perms.rs
@@ -1,4 +1,4 @@
-use crate::commands_v2::prelude::*;
+use crate::commands::prelude::*;
 // use crate::utils::prelude::*;
 
 pub struct Perms;
