+++ src/commands/bot/admin/scheduler.rs
@@ -1,4 +1,4 @@
-use crate::commands_v2::prelude::*;
+use crate::commands::prelude::*;
 // use crate::utils::prelude::*;
 
 pub struct Scheduler;
