+++ src/commands/bot/mod.rs
@@ -32,7 +32,7 @@ impl Command {
 }
 ```
 */
-use crate::commands_v2::{Commands, CommandsBuilder};
+use crate::commands::{Commands, CommandsBuilder};
 use crate::utils::prelude::*;
 
 /// Generic commands.
