+++ src/commands/bot/user/quote.rs
@@ -1,4 +1,4 @@
-use crate::commands_v2::prelude::*;
+use crate::commands::prelude::*;
 // use crate::utils::prelude::*;
 
 /// Command: Quote a random person or manage quotes.
@@ -6,7 +6,7 @@ pub struct Quote;
 
 impl Quote {
     pub fn command() -> impl Into<BaseCommand> {
-        use crate::commands_v2::builder::*;
+        use crate::commands::builder::*;
 
         command("quote", "Get a random quote.")
             .attach(Self::classic)
