+++ src/commands/bot/user/user_info.rs
@@ -1,7 +1,7 @@
 use twilight_mention::Mention;
 use twilight_util::builder::embed::{self, EmbedFieldBuilder, ImageSource};
 
-use crate::commands_v2::prelude::*;
+use crate::commands::prelude::*;
 use crate::utils::prelude::*;
 
 // Useful: https://discord.com/developers/docs/reference#image-formatting-cdn-endpoints
@@ -13,7 +13,7 @@ pub struct UserInfo {
 
 impl UserInfo {
     pub fn command() -> impl Into<BaseCommand> {
-        use crate::commands_v2::builder::*;
+        use crate::commands::builder::*;
 
         command("userinfo", "Get information about a user.")
             .attach(Self::slash)
